refactor: acropolis codec to be modular
