+++ codec/src/lib.rs
@@ -1,2 +1,15 @@
+pub mod address;
 pub mod block;
-pub mod map_parameters;
+pub mod certs;
+pub mod parameter;
+pub mod tx;
+pub mod utils;
+pub mod utxo;
+
+pub use address::*;
+pub use block::*;
+pub use certs::*;
+pub use parameter::*;
+pub use tx::*;
+pub use utils::*;
+pub use utxo::*;
